+++ mulog-kafka/src/com/brunobonacci/mulog/publishers/kafka.clj
@@ -3,19 +3,21 @@
             [com.brunobonacci.mulog.buffer :as rb]
             [com.brunobonacci.mulog.utils :as ut]
             [com.brunobonacci.mulog.common.json :as json]
-            [clojure.string :as str])
+            [clojure.string :as str]
+            [taoensso.nippy :as nippy])
   (:import [java.util Map]
            [org.apache.kafka.clients.producer KafkaProducer ProducerRecord RecordMetadata]
-           [org.apache.kafka.common.serialization StringSerializer Serializer]))
-
+           [org.apache.kafka.common.serialization StringSerializer Serializer
+            ByteArraySerializer]))
 
 
 
 (defn- normalize-config
-  [config]
+  [{:keys [format] :as config}]
   (->> config
+    :kafka
     (merge {:key.serializer   StringSerializer
-            :value.serializer StringSerializer})
+            :value.serializer (if (= :nippy format) ByteArraySerializer StringSerializer)})
     (map (fn [[k v]]
            [(name k)
             (cond
@@ -56,11 +58,11 @@
 
 (comment
 
-  (def kcfg {:bootstrap.servers "192.168.200.200:9092"
+  (def kcfg {:bootstrap.servers "localhost:9092"
              :key.serializer    StringSerializer
              :value.serializer  StringSerializer})
 
-  (def kp (producer kcfg))
+  (def kp (producer {:kafka kcfg}))
 
   (RecordMetadata->data @(send! kp "mulog" "key1" "value1"))
 
@@ -70,12 +72,11 @@
 
 ;; TODO: handle records which can't be serialized.
 (defn- publish-records!
-  [{:keys [key-field format topic producer*] :as  config} records]
-  (let [fmt* (if (= :json format) json/to-json ut/edn-str)]
-    (->> records
-      (map (juxt #(some-> (get % key-field) str) fmt*))
-      (map (fn [[k v]] (send! producer* topic k v)))
-      (doall))))
+  [{:keys [key-field format topic producer* fmt*] :as  config} records]
+  (->> records
+    (map (juxt #(some-> (get % key-field) str) fmt*))
+    (map (fn [[k v]] (send! producer* topic k v)))
+    (doall)))
 
 
 
@@ -123,7 +124,7 @@
 
 
 
-(def ^:const DEFAULT-CONFIG
+(def DEFAULT-CONFIG
   {:max-items     1000
    :publish-delay 1000
    :kafka         {;; the comma-separated list of brokers to connect
@@ -131,12 +132,25 @@
                    ;; you can add more kafka connection properties here
                    }
    :topic         "mulog"
-   ;; one of: :json, :edn
+   ;; one of: :json, :edn, :nippy
    :format        :json
+   ;; nippy configuration
+   :nippy         {:compressor nippy/lz4-compressor}
+   ;; kafka records key
    :key-field     :mulog/trace-id
    ;; function to transform records
    :transform     identity
    })
+
+
+
+(defn- serialization-format
+  [{:keys [format nippy] :as config}]
+  (case format
+    :json json/to-json
+    :edn ut/edn-str
+    :nippy #(nippy/freeze % nippy)
+    json/to-json))
 
 
 
@@ -146,6 +160,7 @@
   (KafkaPublisher.
     (as-> config $
       (ut/deep-merge DEFAULT-CONFIG $)
-      (assoc $ :producer* (producer (:kafka $))))
+      (assoc $ :producer* (producer $))
+      (assoc $ :fmt* (serialization-format $)))
     (rb/agent-buffer 10000)
     (or (:transform config) identity)))